split code_800AF8B0 into 3 files; move rodata into files/asm
