+++ asm/non_matchings/code_800AF9B0/func_800B0350.s
@@ -1,3 +1,11 @@
+.section .late_rodata
+glabel jpt_800F2C20
+.word L800B04B8, L800B04D0, L800B04E8, L800B0500
+.word L800B0510, L800B043C, L800B0454, L800B046C
+.word L800B0484, L800B049C, L800B04B8, L800B04D0
+.word L800B04E8, L800B0500
+
+.section .text
 glabel func_800B0350
 /* 0B0F50 800B0350 3C0E800F */  lui   $t6, %hi(D_800E86A4) # $t6, 0x800f
 /* 0B0F54 800B0354 8DCE86A4 */  lw    $t6, %lo(D_800E86A4)($t6)
