+++ asm/non_matchings/code_800AF9B0/func_800B053C.s
@@ -1,3 +1,57 @@
+.section .late_rodata
+glabel jpt_800F2C58
+.word L800B05E0, L800B05E0
+.word L800B05E0, L800B05E0, L800B13A0, L800B13A0
+.word L800B13A0, L800B13A0, L800B13A0, L800B0AF4
+.word L800B0AF4, L800B0C88, L800B13A0, L800B13A0
+.word L800B13A0, L800B13A0, L800B13A0, L800B13A0
+.word L800B13A0, L800B0CAC, L800B0CAC, L800B1020
+.word L800B1020, L800B1020, L800B1020, L800B13A0
+.word L800B13A0, L800B13A0, L800B13A0, L800B0E5C
+.word L800B0E5C, L800B1020, L800B1020, L800B13A0
+.word L800B1020, L800B1044, L800B1044, L800B11C8
+.word L800B11FC, L800B1020, L800B13A0, L800B13A0
+.word L800B13A0, L800B13A0, L800B1020, L800B1020
+.word L800B13A0, L800B13A0, L800B13A0, L800B1318
+.word L800B134C
+
+.word 0x00000000
+
+glabel D_800F2D28
+.double 4.2
+
+glabel D_800F2D30
+.double 4.2
+
+glabel jpt_800F2D38 # jump table
+.word L800B0948, L800B096C, L800B0980, L800B09C0
+.word L800B09C0, L800B09A0
+
+glabel D_800F2D50
+.double 4.2
+
+glabel D_800F2D58
+.double 4.2
+
+glabel D_800F2D60
+.double 4.2
+
+glabel D_800F2D68
+.double 4.2
+
+glabel D_800F2D70
+.double 4.2
+
+glabel D_800F2D78
+.double 4.2
+
+glabel D_800F2D80
+.double 4.2
+
+glabel D_800F2D88
+.double 4.2
+
+.section .text
 glabel func_800B053C
 /* 0B113C 800B053C 27BDFFC0 */  addiu $sp, $sp, -0x40
 /* 0B1140 800B0540 AFBF001C */  sw    $ra, 0x1c($sp)
