+++ include/macros.h
@@ -62,6 +62,6 @@
 #define VIRTUAL_TO_PHYSICAL2(addr)  ((u8 *)(addr) - 0x80000000U)
 
 // aligns an address to the next 16 bytes
-#define ALIGN16(x) (((x) + 0xF) & -0x10)
+#define ALIGN16(val) (((val) + 0xF) & ~0xF)
 
 #endif
