+++ mk64.ld
@@ -88,6 +88,8 @@ SECTIONS
       BUILD_DIR/src/code_8008C1D0.o(.text);
       BUILD_DIR/src/code_80091750.o(.text);
       BUILD_DIR/src/code_800AF9B0.o(.text);
+      BUILD_DIR/src/code_800B0350.o(.text);
+      BUILD_DIR/src/code_800B45E0.o(.text);
       BUILD_DIR/src/audio/synthesis.o(.text);
       BUILD_DIR/src/audio/heap.o(.text);
       BUILD_DIR/src/audio/load.o(.text);
@@ -250,6 +252,9 @@ SECTIONS
       BUILD_DIR/data/data_0DD0A0_3_0_1.o(.rodata);
       BUILD_DIR/src/code_80057C60.o(.rodata);
       BUILD_DIR/data/data_0DD0A0_3_0.o(.rodata);
+      BUILD_DIR/src/code_800B0350.o(.rodata);
+      BUILD_DIR/src/code_800B45E0.o(.rodata);
+      BUILD_DIR/data/rodata_audio.o(.rodata);
       BUILD_DIR/src/os/guRotateF.o(.rodata);
       BUILD_DIR/src/os/guPerspectiveF.o(.rodata);
       BUILD_DIR/data/data_0DD0A0_3_1.o(.rodata);
