+++ src/code_800029B0.c
@@ -47,7 +47,6 @@ extern u32 D_8015F730;
 extern u32 D_8015F734;
 extern u32 gPrevLoadedAddress;
 extern s32 D_80150120;
-extern s16 gCupCourseOrder[][4];
 
 void func_800CA008(u8, u8);
