+++ src/code_80091750.h
@@ -68,7 +68,9 @@ void func_80092630();
 void func_8009265C();
 void func_80092C80();
 s32  char_to_glyph_index(char*);
-s32  func_80092DF8();
+s32  func_80092DF8(s8 *);
+s32  func_80092E1C(s8 *);
+s32  func_80092EE4(s8 *);
 s32  get_string_width(char*);
 void set_text_color(s32);
 void func_800930E4(s32, s32, s32*);
@@ -113,6 +115,7 @@ void func_8009CA2C();
 void func_8009CA6C(s32);
 void func_80099A94(MkTexture *, s32);
 void func_80099AEC();
+void func_8009E1E4(void);
 void func_80099E54();
 void func_80099E60(MkTexture *, s32, s32);
 void func_80099EC4();
