+++ src/staff_ghosts.h
@@ -1,0 +1,12 @@
+#ifndef STAFF_GHOSTS_H
+#define STAFF_GHOSTS_H
+
+#include "PR/os.h"
+#include "common_structs.h"
+#include <types.h>
+
+s32  func_800051C4();
+void func_8000522C();
+void func_80005AE8(Player *);
+
+#endif /* STAFF_GHOSTS_H */
